requirements.txt
bcrypt==4.2.1
blinker==1.9.0
Brotli==1.1.0
click==8.1.8
csscompressor==0.9.5
cssmin==0.2.0
Flask==3.1.0
Flask-Assets==2.1.0
Flask-Bcrypt==1.0.1
Flask-Compress==1.17
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsmin==3.0.1
MarkupSafe==3.0.2
pillow==11.1.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
typing_extensions==4.12.2
webassets==2.0
Werkzeug==3.1.3
WTForms==3.2.1
zstandard==0.23.0
